+++ Configuration/Icons.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    // Icon identifier
+    'web-popuppower-dashboard' => [
+        // Icon provider class
+        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        // The source SVG for the SvgIconProvider
+        'source' => 'EXT:popup_power/Resources/Public/Icons/Extension.svg',
+    ],
+];