+++ ext_tables.sql
@@ -1,0 +1,5 @@
+CREATE TABLE tx_popuppower_domain_model_configuration(
+	`hidden` TINYINT DEFAULT '0' NOT NULL,
+	`name` varchar(256) DEFAULT 'Default' NOT NULL,
+	`extend_to_subpages` TINYINT DEFAULT '0' NOT NULL
+);