+++ CHANGES.md
@@ -1,3 +1,7 @@
+### Version 0.0.2
+- [IMPORTANT] Add Backend Module to create and remove configuration
+- [FEATURE] Configuration model to enable popup for pages
+
 ### Version 0.0.1
 - [IMPORTANT] Kickstart extension
 - [FEATURE] Setup GitHub workflow for CGL V12 and V13